// frontend/src/types/movie.ts
export type Movie = {
  id: number;
  title: string;
  overview: string;
  poster_path: string | null;
  backdrop_path: string | null;
  release_date: string;
  vote_average: number;
  genre_ids: number[];
};

export type MovieResponse = { results: Movie[]; page: number; total_pages: number; total_results: number };